+++ kifi-backend/modules/abook/app/com/keepit/dev/ABookDevModule.scala
@@ -7,7 +7,7 @@ import com.keepit.inject.ProdFortyTwoModule
 import com.keepit.common.actor.DevActorSystemModule
 import com.keepit.common.zookeeper.DevDiscoveryModule
 import com.keepit.social.RemoteSecureSocialModule
-import com.keepit.abook.ABookModule
+import com.keepit.abook.{DevContactsUpdaterPluginModule, ProdContactsUpdaterPluginModule, ABookModule}
 import com.keepit.common.healthcheck.ProdHealthCheckModule
 import com.keepit.common.cache.HashMapMemoryCacheModule
 import com.keepit.inject.CommonDevModule
@@ -15,7 +15,8 @@ import com.keepit.common.store.ABookDevStoreModule
 
 case class ABookDevModule() extends ABookModule (
   cacheModule = ABookCacheModule(HashMapMemoryCacheModule()),
-  storeModule = ABookDevStoreModule()
+  storeModule = ABookDevStoreModule(),
+  contactsUpdaterPluginModule = DevContactsUpdaterPluginModule()
 ) with CommonDevModule {
 
 }
